+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This list is not currently intended to be all-encompassing - it will document ma
 changes with their rationale when appropriate:
 
 ### v3.257.0
-- [http4k-serverless-*] Making the APIs consistent between flavours by ensu
+- [http4k-serverless-*] Making the Serverless APIs consistent between flavours by ensuring that all Serverless functions act by class extension and not reflection based approach. Deprecated old approach. Hopefully this is simpler.. :) 
 
 ### v3.256.1
 - [http4k-core] Fix #470. Path.of cannot decode path parameter values containing %/